+++ client/design/designSurface/drawPathsTool/GuidePath.js
@@ -1,9 +1,8 @@
-import FieldDimensions from '/client/lib/FieldDimensions';
 import StrideType from '/client/lib/StrideType';
 import StepType from '/client/lib/StepType';
 import StepDelta from '/client/lib/StepDelta';
 import Direction from '/client/lib/Direction';
-import { FieldPoint, StepPoint } from '/client/lib/Point';
+import { StepPoint } from '/client/lib/Point';
 import TurnMarker from '../field/TurnMarker';
 import CounterMarch from '../field/CounterMarch';
 import PathUtils from './PathUtils';
@@ -17,7 +16,7 @@ class GuidePath {
             direction: initialPoint.direction,
             x: initialPoint.x,
             y: initialPoint.y,
-        }; 
+        };
         this.strideType = strideType;
         this.points = [this.initialPoint];
         this.file = file;
@@ -40,46 +39,48 @@ class GuidePath {
     }
 
     findPoint(point) {
-        var foundPoint = this.points.find(p => {
+        let foundPoint = this.points.find((p) => {
             return p.x == point.x && p.y == point.y;
         });
         return foundPoint;
     }
 
     findPrecedingPoint(point) {
-        var i = this.points.indexOf(point);
+        let i = this.points.indexOf(point);
         return this.points[i - 1];
     }
 
     // return newPoint if it is a valid point, orig if not
     movePoint(origPoint, newPoint) {
         // var p = this.findPoint(origPoint);
-        // if (!p) 
+        // if (!p)
         //     return this.resetMarkerPoint(p.turnMarker, p);
 
-        var prevPoint = this.findPrecedingPoint(origPoint);
-        if (!this.isInPath(newPoint, prevPoint))
+        let prevPoint = this.findPrecedingPoint(origPoint);
+        if (!this.isInPath(newPoint, prevPoint)) {
             return this.resetMarkerPoint(origPoint.turnMarker, origPoint);
-        
-        var delta = {
+        }
+
+        let delta = {
             deltaX: newPoint.x - origPoint.x,
-            deltaY: newPoint.y - origPoint.y
+            deltaY: newPoint.y - origPoint.y,
         };
 
         // adjust point and those that follow it
-        for (var i = this.points.indexOf(origPoint); i < this.points.length; i++) {
+        for (let i = this.points.indexOf(origPoint); i < this.points.length; i++) { // eslint-disable-line max-len
             let p2 = this.points[i];
             p2.x = p2.x + delta.deltaX;
             p2.y = p2.y + delta.deltaY;
 
-            var newFp = new StepPoint(p2.strideType, p2.x, p2.y).toFieldPoint();
+            let newFp = new StepPoint(p2.strideType, p2.x, p2.y).toFieldPoint();
             this.setMarkerPoint(p2.turnMarker, newFp);
         }
     }
 
     resetMarkerPoint(marker, point) {
         if (!marker) return;
-        var fp = new StepPoint(point.strideType, point.x, point.y).toFieldPoint();
+        let fp = new StepPoint(point.strideType, point.x, point.y)
+            .toFieldPoint();
         this.setMarkerPoint(marker, fp);
     }
 
@@ -92,92 +93,104 @@ class GuidePath {
     }
 
     add(point) {
-        var snappedPoint = PathUtils.snapPoint(this.strideType, this.lastPoint, point);
+        let snappedPoint = PathUtils.snapPoint(this.strideType,
+            this.lastPoint, point);
         if (snappedPoint.direction === Direction.CM) {
             this.addCountermarch(snappedPoint);
             return;
         }
-        snappedPoint.stepsFromPrevious = this.calculateStepsFromPreviousPoint(snappedPoint);
+        snappedPoint.stepsFromPrevious = this
+            .calculateStepsFromPreviousPoint(snappedPoint);
         this.createTurnMarker(snappedPoint);
         this.points.push(snappedPoint);
         this.createGuidePathLine();
     }
 
     isLeftTurn(stepPoint) {
-        return this.getEndCount(stepPoint) % 2 == 0 ? true : false;        
+        return this.getEndCount(stepPoint) % 2 == 0 ? true : false;
     }
 
     addCountermarch(point) {
-        var stepPoint = new StepPoint(point.strideType, point.x, point.y);
-        var isLeftTurn = this.isLeftTurn(stepPoint);
+        let stepPoint = new StepPoint(point.strideType, point.x, point.y);
+        let isLeftTurn = this.isLeftTurn(stepPoint);
 
-        var currentDir = this.lastPoint.direction;
-        var firstTurnDirection = isLeftTurn ? Direction.leftTurnDirection(currentDir) : Direction.rightTurnDirection(currentDir);
-        var secondTurnDirection = isLeftTurn ? Direction.leftTurnDirection(firstTurnDirection) : Direction.rightTurnDirection(firstTurnDirection);
-        var firstDelta = StepDelta.getDelta(this.strideType, StepType.Half, firstTurnDirection, 2);
+        let currentDir = this.lastPoint.direction;
+        let firstTurnDirection = isLeftTurn
+            ? Direction.leftTurnDirection(currentDir)
+            : Direction.rightTurnDirection(currentDir);
+        let secondTurnDirection = isLeftTurn
+            ? Direction.leftTurnDirection(firstTurnDirection)
+            : Direction.rightTurnDirection(firstTurnDirection);
+        let firstDelta = StepDelta.getDelta(this.strideType,
+            StepType.Half, firstTurnDirection, 2);
 
-        var fp = stepPoint.toFieldPoint();
-        var tm = new CounterMarch(currentDir, isLeftTurn, {
+        let fp = stepPoint.toFieldPoint();
+        let tm = new CounterMarch(currentDir, isLeftTurn, {
             left: fp.x,
-            top: fp.y
+            top: fp.y,
         });
         this.field.canvas.add(tm);
 
-        var point1 = {
+        let point1 = {
             x: stepPoint.x,
             y: stepPoint.y,
             direction: firstTurnDirection,
             strideType: this.strideType,
             stepType: StepType.Half,
-            turnMarker: tm
+            turnMarker: tm,
         };
         point1.stepsFromPrevious = this.calculateStepsFromPreviousPoint(point1);
         this.points.push(point1);
 
-        var secondDelta = StepDelta.getDelta(this.strideType, StepType.Full, secondTurnDirection, 1);
+//        let secondDelta = StepDelta.getDelta(this.strideType, StepType.Full, secondTurnDirection, 1);
 
-        var point2 = {
+        let point2 = {
             x: stepPoint.x + firstDelta.deltaX,
             y: stepPoint.y + firstDelta.deltaY,
             direction: secondTurnDirection,
             strideType: this.strideType,
             stepType: StepType.Full,
-            stepsFromPrevious: 2
-            //turnMarker: tm  // deliberately don't reference tm here, interferes with moving
+            stepsFromPrevious: 2,
+            // turnMarker: tm  // deliberately don't reference tm here, interferes with moving
         };
         this.points.push(point2);
         tm.point = point1;
         tm.point1 = point1;
         tm.point2 = point2;
 
-        tm.on('moving', evt => { this.onMoveTurnMarker(evt, this.field, this, point, tm); });
-        
+        tm.on('moving', (evt) => {
+this.onMoveTurnMarker(evt, this.field, this, point, tm);
+});
+
         this.createGuidePathLine();
     }
 
     calculateStepsFromPreviousPoint(point) {
-        var prevPoint = this.lastPoint;
-        return StepDelta.getStepsBetweenPoints(this.strideType || StrideType.SixToFive, prevPoint.stepType || StepType.Full, prevPoint.direction, point, prevPoint);
+        let prevPoint = this.lastPoint;
+        return StepDelta.getStepsBetweenPoints(this.strideType
+            || StrideType.SixToFive, prevPoint.stepType
+            || StepType.Full, prevPoint.direction, point, prevPoint);
     }
 
     createTurnMarker(point) {
         point.turnMarker = new TurnMarker(point.direction, {
             left: point.x,
-            top: point.y
+            top: point.y,
         });
         point.turnMarker.point = point;
-        point.turnMarker.on('moving', evt => { 
-            this.onMoveTurnMarker(evt, this.field, this, point, point.turnMarker); 
+        point.turnMarker.on('moving', (evt) => {
+            this.onMoveTurnMarker(evt, this.field, this,
+                point, point.turnMarker);
             this.turnMarkerMoving = true;
         });
-        point.turnMarker.on('modified', evt => {
+        point.turnMarker.on('modified', (evt) => {
             this.turnMarkerMoving = false;
         });
         this.field.canvas.add(point.turnMarker);
     }
 
     removeTurnMarker(marker) {
-        if(marker.type == 'CounterMarch') {
+        if (marker.type == 'CounterMarch') {
             return this.removeCounterMarch(marker);
         }
 
@@ -200,24 +213,26 @@ class GuidePath {
     }
 
     removePoint(p) {
-        var pIndex = this.points.indexOf(p);
+        let pIndex = this.points.indexOf(p);
         if (pIndex < 0) return;
 
         // get delta between point being removed and following point
-        var nextPoint = pIndex < this.points.length - 1 ? this.points[pIndex + 1] : { x: 0, y: 0 };
-        var delta = {
+        let nextPoint = pIndex < this.points.length - 1
+            ? this.points[pIndex + 1]
+            : { x: 0, y: 0 };
+        let delta = {
             deltaX: p.x - nextPoint.x,
-            deltaY: p.y - nextPoint.y
+            deltaY: p.y - nextPoint.y,
         };
 
         // adjust point and those that follow it
-        for (var i = pIndex + 1; i < this.points.length; i++) {
+        for (let i = pIndex + 1; i < this.points.length; i++) {
             let p2 = this.points[i];
-            
+
             p2.x = p2.x + delta.deltaX;
             p2.y = p2.y + delta.deltaY;
 
-            var newFp = new StepPoint(p2.strideType, p2.x, p2.y).toFieldPoint();
+            let newFp = new StepPoint(p2.strideType, p2.x, p2.y).toFieldPoint();
             this.setMarkerPoint(p2.turnMarker, newFp);
         }
 
@@ -226,27 +241,31 @@ class GuidePath {
     }
 
     bringTurnMarkersToFront() {
-        this.points.forEach(p => {
+        this.points.forEach((p) => {
             this.field.canvas.bringToFront(p.turnMarker);
         });
     }
 
 
     onMouseMove(evt) {
-        if( this.turnMarkerMoving) return;
-        var adjustedPoint = this.field.adjustMousePoint({ x: evt.e.layerX, y: evt.e.layerY });
-        //var stepPoint = new FieldPoint(adjustedPoint).toStepPoint(this.strideType);
+        if ( this.turnMarkerMoving) return;
+        let adjustedPoint = this.field.adjustMousePoint({
+            x: evt.e.layerX,
+            y: evt.e.layerY,
+        });
+        // var stepPoint = new FieldPoint(adjustedPoint).toStepPoint(this.strideType);
 this.shiftKey = evt.e.shiftKey;
         if (this.isInPath(adjustedPoint)) {
-            var snappedPoint = PathUtils.snapPoint(this.strideType, this.lastPoint, adjustedPoint);
+            let snappedPoint = PathUtils.snapPoint(this.strideType,
+                this.lastPoint, adjustedPoint);
 //            snappedPoint.steps = this.calculateStepsFromPreviousPoint(snappedPoint);
             this.createGuideline(this.lastPoint, snappedPoint);
-            
+
             if (this.currentDir == Direction.CM) {
                 if (this.isLeftTurn(snappedPoint)) {
-                    this.field.canvas.defaultCursor = "url(/icons/CM-left.svg) 8 8, auto";
+                    this.field.canvas.defaultCursor = 'url(/icons/CM-left.svg) 8 8, auto';
                 } else {
-                    this.field.canvas.defaultCursor = "url(/icons/CM-right.svg) 8 8, auto";
+                    this.field.canvas.defaultCursor = 'url(/icons/CM-right.svg) 8 8, auto';
                 }
             }
         } else {
@@ -255,10 +274,14 @@ this.shiftKey = evt.e.shiftKey;
     }
 
     onMoveTurnMarker(evt, field, guidePath, point, turnMarker) {
-        let adjustedMousePoint = field.adjustMousePoint({ x: evt.e.layerX, y: evt.e.layerY });
+        let adjustedMousePoint = field.adjustMousePoint({
+            x: evt.e.layerX,
+            y: evt.e.layerY,
+        });
         let precedingPoint = this.findPrecedingPoint(turnMarker.point);
-        let snappedPoint = PathUtils.snapPoint(this.strideType, precedingPoint, adjustedMousePoint);
-        //let moveToStepPoint = new FieldPoint(adjustedMousePoint).toStepPoint(this.strideType);
+        let snappedPoint = PathUtils.snapPoint(this.strideType,
+            precedingPoint, adjustedMousePoint);
+        // let moveToStepPoint = new FieldPoint(adjustedMousePoint).toStepPoint(this.strideType);
 
         // this.setGuidelineLabel('foo');
         // this.createGuideline(precedingPoint, moveToStepPoint);
@@ -277,7 +300,7 @@ this.shiftKey = evt.e.shiftKey;
             strokeDashArray: [3, 3],
             fill: false,
             selectable: false,
-            evented: false
+            evented: false,
         });
         this.field.canvas.add(this.path);
         this.bringTurnMarkersToFront();
@@ -298,14 +321,14 @@ this.shiftKey = evt.e.shiftKey;
             strokeWidth: 2,
             strokeDashArray: [2, 2],
             selectable: false,
-            evented: false
+            evented: false,
         });
 
-        var steps = to.steps || 0;
+        let steps = to.steps || 0;
 
         function formatSteps(x) {
             if (x == 0) return '';
-            return x.toFixed(1).replace(/\.0$/, '');     
+            return x.toFixed(1).replace(/\.0$/, '');
         }
 
         this.guidelineLabel = new fabric.Text(formatSteps(steps), {
@@ -316,9 +339,9 @@ this.shiftKey = evt.e.shiftKey;
 //            fontWeight: 'bold',
 //            lineHeight: 1,
             selectable: false,
-            evented: false
+            evented: false,
           });
-          
+
         this.field.canvas.add(this.guideline);
         this.field.canvas.add(this.guidelineLabel);
         this.bringTurnMarkersToFront();
@@ -332,7 +355,7 @@ this.shiftKey = evt.e.shiftKey;
 
     createMoveStepsLabel(from, to) {
         this.destroyMoveStepsLabel();
-        var steps = (to.steps || 0).toString();
+        let steps = (to.steps || 0).toString();
         this.moveStepsLabel = new fabric.Text(steps, {
             left: to.x + 10,
             top: to.y + 10,
@@ -341,10 +364,10 @@ this.shiftKey = evt.e.shiftKey;
 //            fontWeight: 'bold',
 //            lineHeight: 1,
             selectable: false,
-            evented: false
+            evented: false,
           });
-          
-        this.field.canvas.add(this.moveStepsLabel);        
+
+        this.field.canvas.add(this.moveStepsLabel);
     }
 
     destroyMoveStepsLabel() {
@@ -363,28 +386,33 @@ this.shiftKey = evt.e.shiftKey;
     }
 
     getEndCount(stepPoint) {
-        if (!stepPoint)
+        if (!stepPoint) {
             return this.startCount + this.getLengthInCounts();
+        }
 
-        var steps = StepDelta.getStepsBetweenPoints(this.strideType, stepPoint.stepType, this.lastPoint.direction, this.lastPoint, stepPoint);
+        let steps = StepDelta.getStepsBetweenPoints(this.strideType,
+            stepPoint.stepType, this.lastPoint.direction,
+            this.lastPoint, stepPoint);
 
         return this.startCount + this.getLengthInCounts() + steps;
     }
 
     getLengthInCounts() {
-        var counts = 0;
-        if (this.points.length < 2)
+        let counts = 0;
+        if (this.points.length < 2) {
             return 0;
-        for (var i = 0; i < this.points.length - 1; i++) {
+        }
+        for (let i = 0; i < this.points.length - 1; i++) {
             let a = this.points[i];
             let b = this.points[i + 1];
-            counts += StepDelta.getStepsBetweenPoints(this.strideType, a.stepType, a.direction, a, b);
+            counts += StepDelta.getStepsBetweenPoints(this.strideType,
+                a.stepType, a.direction, a, b);
         }
         return counts;
     }
 
     isInPath(point, sourcePoint) {
-        var sourcePoint = sourcePoint || this.lastPoint;
+        sourcePoint = sourcePoint || this.lastPoint;
 
         return PathUtils.isInPath(this.strideType, sourcePoint, point);
     }
@@ -395,14 +423,15 @@ this.shiftKey = evt.e.shiftKey;
     }
 
     isOnLine(knownPoint, knownSlope, testPoint) {
-        var b = this.yIntercept(knownPoint, knownSlope);
-        return testPoint.y.toFixed(3) == ((knownSlope * testPoint.x) + b).toFixed(3); // y = mx + b
+        let b = this.yIntercept(knownPoint, knownSlope);
+        return testPoint.y.toFixed(3) == ((knownSlope * testPoint.x) + b)
+            .toFixed(3); // y = mx + b
     }
 
     get pathExpr() {
-        var pathExpr = "M ";
-        this.points.forEach(sp => {
-            //let fp = new StepPoint(sp.strideType || StrideType.SixToFive, sp.x, sp.y).toFieldPoint();
+        let pathExpr = 'M ';
+        this.points.forEach((sp) => {
+            // let fp = new StepPoint(sp.strideType || StrideType.SixToFive, sp.x, sp.y).toFieldPoint();
             pathExpr += sp.x + ' ' + sp.y + ' L ';
         });
         return pathExpr.slice(0, -2);
@@ -412,7 +441,7 @@ this.shiftKey = evt.e.shiftKey;
         this.field.canvas.off('mouse:move', this.onMouseMoveHandler);
         this.destroyGuideline();
         this.destroyGuidePathLine();
-        this.points.forEach(p => {
+        this.points.forEach((p) => {
             this.field.canvas.remove(p.turnMarker);
             p.turnMarker = null;
         });